Extend export_images function to allow image splitting
Added two optional parameters: 'split_lr' or 'flip_lr'.
If 'split_lr == true' the image is split into a left and a right half and two files per example are written instead of one.
If 'flip_lr == True' the right image is flipped vertically, so it appears like a left image.